Added a new root page
We have creted a new root page to avoid problems deploying in Heroku.
For some reason de default rails root page doesn't work fine. So you
have to create a simple one by implementing an action method in the
application controller and setting a root route pointing to the new
action method in the routes.rb file in the config folder.